Some moves and renames.
